requirements.txt
langchain
langchain-openai
streamlit
pypdf
langchain_community
python-dotenv
langchain-chroma